rewrite as async library based on `embedded-hal-async` - WIP
this makes it portable and usable on all devices, including embedded
devices without linux.

as part of the upgrade the library was also updated to the rust edition
2024, whcih requires at least rustc 1.85.0. this is not strictly
necessary, the edition 2021 (with a lower rustc requirement) would
suffice, but this can still be downgraded if someone really needs to use
an older compiler with this library.

TODOs:
- manually test on real hardware
- add test coverage
- add add docs for everything